requirements.txt
flask
flask_restful
pymupdf
spacy
psycopg2-binary
openai
scikit-learn